fix(expense): a fractional quantity no longer 422s at the API boundary
The structural twin of the bill_line_item fix (34a4ccfa, 06108842,
27c3da22), applied to entities/expense_line_item. Same defect: the column
is DECIMAL(18,4) and the sprocs take @Quantity DECIMAL(18,4), but Python
typed quantity as int on the write path while repo.py:49 already hydrated
Decimal on the read path — so the two halves of one column disagreed.

Weaker than bill was, and now repaired: service.py:231 was a bare
`existing.quantity = quantity` with no coercion at all, and repo.py:82/:227
bound raw values, sitting directly above four already-coerced siblings.

⚠️ NO DATA BACKFILL IS OWED, and that is the material difference from the
bill side. Measured live: dbo.ExpenseLineItem 12,634 rows / 5,691 with a
quantity / ZERO fractional; qbo.PurchaseLine 5,310 / ZERO fractional, ZERO
over 4dp; range 0–1; no row exceeds the new ceiling. The purchase pull used
Decimal(str(...)) throughout and never int-coerced, so nothing was ever
truncated here. This fix is purely forward-looking — bill needed 55 rows
repaired, expense needs none.

complete_expense does NOT carry the complete_bill ValidationError-swallowing
bug: it passes kwargs rather than constructing ExpenseLineItemUpdate (which
has no referents outside schemas.py/router.py). Pinned behaviourally, so a
future "make it consistent with bill" refactor goes RED.

max_digits=18 AND le/ge at the column ceiling, both independently
mutation-proven RED — deliberately avoiding the lapse 27c3da22 had to fix on
the bill side, where adding le/ge silently un-pinned max_digits because every
over-range case was large-magnitude.

Review corrections folded in, both from the independent pass:
  - The serialisation pin was VACUOUS. `jsonable_encoder` maps Decimal("5.0000")
    -> 5.0, int 5 -> 5 and float 5.0 -> 5.0, all of which compare == 5.0, so
    reverting the read path to int() left it green — it could not detect the
    regression it was named for. Now asserts the JSON TEXT ("5.0" vs an int's
    "5"), and is proven RED against an int-truncating read path.
  - The decimal_places rationale was WRONG. It said the QBO pull would break;
    the pull calls the service directly and never sees this schema, so a scale
    bound could not have blocked it. The honest reason is edge-vs-internal
    consistency: the system behind this API stores 6dp and rounds at write, so
    rejecting 5.016667 at the edge makes the door stricter than the room.

Reviewer: independent subagent, VERDICT PASS, no P0/P1/P2. ⚠️ Codex ran out
of credits mid-review and emitted no verdict — that leg is UNADJUDICATED, not
a pass. Three other subagents died to a watchdog. Bounds were mutation-verified
directly as the compensating control.

Verify: 5246 passed / 2 skipped.

Out of scope, reported not fixed (both twins of the bill residuals):
entities/expense/api/schemas.py:67 line_quantity:int still 422s the
parent-with-line create path, and entities/expense/intelligence/tools.py
:257/:472/:552 keep agent tools on int.

Unit: U-503e

+++ entities/expense_line_item/business/model.py
@@ -20,7 +20,7 @@ class ExpenseLineItem:
     sub_cost_code_id: Optional[int]
     project_id: Optional[int]
     description: Optional[str]
-    quantity: Optional[int]
+    quantity: Optional[Decimal]
     rate: Optional[Decimal]
     amount: Optional[Decimal]
     is_billable: Optional[bool]
